requirements.txt
aioredis==1.3.1
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cookiecutter==1.7.3
cryptography==36.0.2
daphne==3.0.2
Django==3.2.13
django-browser-reload==1.3.0
django-htmx==1.9.0
django-tailwind==3.1.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
msgpack==1.0.3
Pillow==9.1.0
poyo==0.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.1
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
zope.interface==5.4.0
